+++ controllers/universityController.js
@@ -1,10 +1,10 @@
 const University = require('../models/universityModel');
 const Course = require('../models/courseModel');
-
+const mongoose = require('mongoose');
 
 const getUniversities = async (req, res) => {
     try {
-        const universities = await University.find({}).populate('courses'); // Popola i dettagli dei corsi
+        const universities = await University.find({}).populate('courses'); // Add courses details
         res.status(200).json(universities);
     } catch (error) {
         res.status(500).json({message: error.message});
@@ -23,7 +23,13 @@ const addUniversity = async (req, res) => {
 const updateUniversity = async (req, res) => {
     try {
         const { id } = req.params;
-        const updatedUniversity = await University.findByIdAndUpdate(id, req.body, { new: true }); //{filtro}, dati da aggiornare, {opzione per restituire documenti aggiornato}
+
+        // Check if ID is valid
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({ message: "Invalid university ID." });
+        }
+
+        const updatedUniversity = await University.findByIdAndUpdate(id, req.body, { new: true }); //{filter}, datas to add, {option to get updated document}
 
         if(!updatedUniversity) {
             return res.status(404).json({message: "University not found."});
@@ -38,6 +44,12 @@ const updateUniversity = async (req, res) => {
 const deleteUniversity = async (req, res) => {
     try {
         const { id } = req.params;
+
+        // Check if ID is valid
+        if (!mongoose.Types.ObjectId.isValid(id)) {
+            return res.status(400).json({ message: "Invalid university ID." });
+        }
+
         const university = await University.findByIdAndDelete(id);
 
         if(!university) {
